backend/requirements.txt
Django==6.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
